move generic state chain functions to new module
